Consolidate subscriptions module into payments
