+++ squarelet/organizations/views/subscription.py
@@ -51,7 +51,7 @@ from squarelet.organizations.tasks import (
     handle_subscription_deleted,
     handle_subscription_updated,
 )
-from squarelet.subscriptions.views import (
+from squarelet.payments.views import (
     BaseCancelSubscription,
     BaseManageSubscriptions,
     BasePaymentsList,
